Adding compiling version split into plugins - not runable due to library duplication
