+++ GeneratorInterface/ExternalDecays/BuildFile.xml
@@ -1,11 +1,3 @@
-<use   name="FWCore/ParameterSet"/>
-<use   name="GeneratorInterface/Pythia6Interface"/>
-<use   name="heppdt"/>
-<use   name="SimGeneral/HepPDTRecord"/>
-<use   name="tauola"/>
-<use   name="photos"/>
-<use   name="clhep"/>
-<use   name="evtgenlhc"/>
-<export>
-  <lib   name="1"/>
-</export>
+
+
+
