+++ GeneratorInterface/ExternalDecays/plugins/EventGen/EvtGenInterface.cc
@@ -15,6 +15,25 @@
 #include "FWCore/ServiceRegistry/interface/Service.h"
 #include "FWCore/Utilities/interface/RandomNumberGenerator.h"
 #include "CLHEP/Random/Random.h"
+#include "CLHEP/Random/RandFlat.h"
+
+#include "EvtGen/EvtGen.hh"
+#include "EvtGenBase/EvtId.hh"
+#include "EvtGenBase/EvtPDL.hh"
+#include "EvtGenBase/EvtDecayTable.hh"
+#include "EvtGenBase/EvtSpinType.hh"
+#include "EvtGenBase/EvtVector4R.hh"
+#include "EvtGenBase/EvtParticle.hh"
+#include "EvtGenBase/EvtScalarParticle.hh"
+#include "EvtGenBase/EvtStringParticle.hh"
+#include "EvtGenBase/EvtDiracParticle.hh"
+#include "EvtGenBase/EvtVectorParticle.hh"
+#include "EvtGenBase/EvtRaritaSchwingerParticle.hh"
+#include "EvtGenBase/EvtTensorParticle.hh"
+#include "EvtGenBase/EvtHighSpinParticle.hh"
+#include "EvtGenBase/EvtStdHep.hh"
+#include "EvtGenBase/EvtSecondary.hh"
+#include "EvtGenModels/EvtPythia.hh"
 
 #include "GeneratorInterface/ExternalDecays/interface/myEvtRandomEngine.h"
 #include "GeneratorInterface/Pythia6Interface/interface/Pythia6Service.h"
