+++ GeneratorInterface/ExternalDecays/plugins/EventGen/myphotos.f
