+++ GeneratorInterface/ExternalDecays/plugins/EventGen/myphotoshelper.F
